+++ package.json
@@ -39,6 +39,8 @@
     "react-error-overlay": "^1.0.10",
     "react-p5-wrapper": "^0.0.4",
     "react-router-dom": "^4.1.2",
+    "react-router-hash-link": "^1.1.1",
+    "react-scrollchor": "^4.0.0",
     "react-youtube": "^7.4.0",
     "rebass": "^1.0.4",
     "style-loader": "0.18.2",
